+++ sjl-minimal.js
@@ -1,5 +1,5 @@
 /**! 
- * sjl-minimal.js Wed Aug 19 2015 17:16:29 GMT-0400 (Eastern Daylight Time)
+ * sjl-minimal.js Thu Sep 10 2015 15:44:17 GMT-0400 (EDT)
  **/
 /**
  * Created by Ely on 5/29/2015.
@@ -665,10 +665,10 @@
      */
     sjl.getValueFromObj = function (key, obj, args, raw, noLegacyGetters) {
         // Warn user(s) of new update to this function where `raw` is not being passed in.
-        if (typeof raw === 'undefined') {
-            console.warn('`sjl.getValueFromObj` now has it\'s `raw` parameter set to `true` by default.  ' +
-                'This warning will be removed in the next library update.');
-        }
+        //if (typeof raw === 'undefined') {
+        //    console.warn('`sjl.getValueFromObj` now has it\'s `raw` parameter set to `true` by default.  ' +
+        //        'This warning will be removed in the next library update.');
+        //}
         args = args || null;
         raw = raw || true;
         noLegacyGetters = typeof noLegacyGetters === 'undefined' ? false : noLegacyGetters;